Compute rainfall erosivity from an hyetograph

hyet_erosivity uses an hyetograph, single or grouped (i.e. from hyet_split) and computes the rainfall erosivity.

hyet_erosivity(hyet, time_step, en_equation = "brown_foster")

Arguments

hyet
an hyetograph from hyet_create function. Precipitation values must be in millimeters (mm).

time_step
hyetograph's time-step in minutes. Valid value is on of [5, 10, 15, 30].

en_equation
a character string specifying the equation to be used for calculating kinetic energy of rainfall. Must have one of the values: "brown_foster", "mcgregor_mutch" or "wisch_smith". Default value is "brown_foster".

The rainfall's kinetic energy equations are:

Brown and Foster (1987), \(e = 0.29(1 - 0.72 e^{-0.05i})\)

McGregor and Mutchler (1976), \(e = 0.273 + 0.2168e^{-0.048i} - 0.4126e^{-0.072i}\)

Wischmeier and Smith (1958), \(e = 0.119 + 0.0873log_{10}(i)\), with the upper limit of 0.283 MJ/ha/mm if \(i\) > 76 mm/h.

In the above equations \(i\) is rainfall intensity (mm/hr) and \(e\) is the kinetic energy per unit of rainfall (MJ/ha/mm).

Value

a tibble with erosivity events related values: staring and ending date, cumulative precipitation, maximum rolling intensity values of 15 and 30 minutes, total energy, erosivity and erosivity density. Only the events with cumulative precipitation greater than 1.27 mm and duration greater than 30 minutes are returned.

Note

hyet must not contain missing dates. Please use the hyet_fill function before using hyet_erosivity.

References

Brown, L. C., & Foster, G. R. (1987). Storm erosivity using idealized intensity distributions. Transactions of the ASAE, 30(2), 379-0386.

McGregor, K. C., Bingner, R. L., Bowie, A. J., & Foster, G. R. (1995). Erosivity index values for northern Mississippi. Transactions of the ASAE, 38(4), 1039-1047.

Wischmeier, W. H., & Smith, D. D. (1958). Rainfall energy and its relationship to soil loss. Eos, Transactions American Geophysical Union, 39(2), 285-291.

Examples

# create time series with 30 mins time step
time_step <- 30
len <- 12
en_equation <- "brown_foster"

hyet <- tibble::tibble(
  date = seq(
    from = as.POSIXct(0, origin = "2018-01-01 00:00:00", tz = "UTC"),
    length.out = len,
    by = paste(time_step, "mins")
  ),
  prec = c(1.1, 2.3, 3.2, 1.9, 4.1, 5.9, 2.5, 3.1, 2.9, 1.2, 0.5, 0.2)
)

# compute erosivity
hyet %>%
hyet_erosivity(time_step, en_equation)

#> # A tibble: 1 x 9
#>   begin               end                 duration cum_prec max_i15 max_i30
#>   <dttm>              <dttm>              <time>      <dbl>   <dbl>   <dbl>
#> 1 2018-01-01 00:00:00 2018-01-01 05:30:00 360 mins     28.9       0    11.8
#> # ... with 3 more variables: total_energy <dbl>, erosivity <dbl>,
#> #   eros_density <dbl>